[AIP-21] Move contrib.operators.s3_to_sftp_operator providers.amazon.aws.operators.s3_to_sftp
