+++ scripts/ci/pylint_todo.txt
@@ -53,7 +53,7 @@
 ./airflow/providers/amazon/aws/operators/s3_delete_objects.py
 ./airflow/providers/amazon/aws/operators/s3_list.py
 ./airflow/contrib/operators/s3_to_gcs_operator.py
-./airflow/contrib/operators/s3_to_sftp_operator.py
+./airflow/providers/amazon/aws/operators/s3_to_sftp.py
 ./airflow/providers/amazon/aws/operators/sagemaker_base.py
 ./airflow/providers/amazon/aws/operators/sagemaker_endpoint_config.py
 ./airflow/providers/amazon/aws/operators/sagemaker_endpoint.py
