+++ tests/providers/amazon/aws/operators/test_s3_to_sftp.py
@@ -23,8 +23,8 @@ import boto3
 from moto import mock_s3
 
 from airflow.configuration import conf
-from airflow.contrib.operators.s3_to_sftp_operator import S3ToSFTPOperator
 from airflow.models import DAG, TaskInstance
+from airflow.providers.amazon.aws.operators.s3_to_sftp import S3ToSFTPOperator
 from airflow.providers.ssh.operators.ssh import SSHOperator
 from airflow.utils import timezone
 from airflow.utils.timezone import datetime
